renames package dataset => datasets
