+++ src/dimcat/data/datasets/base.py
