+++ src/dimcat/steps/analyzers/base.py
@@ -7,7 +7,7 @@ from typing import Any, ClassVar, Iterable, Optional, Type, TypeVar
 
 import marshmallow as mm
 from dimcat.base import ObjectEnum
-from dimcat.data.dataset.processed import AnalyzedDataset
+from dimcat.data.datasets.processed import AnalyzedDataset
 from dimcat.data.resource.base import SomeSeries
 from dimcat.data.resource.dc import DimcatResource
 from dimcat.data.resource.features import Feature, FeatureSpecs
